Prepare to rewrite overlapping instances
This uses the technique that is used in the Show class
to make show for String use quotes instead of brackets.

This prepares to remove the overlap in Val [a] and Val (Maybe a).

Move val and cast' in Val [a] to valList and cast'List in Val a.

Move val and cast' in Val (Maybe a) to valMaybe and cast'Maybe in Val a.